+++ packages/client/src/subscribe/waitForStorage.ts
@@ -1,6 +1,6 @@
-import { toStreamPartID } from '@streamr/protocol'
+import { StreamMessage, toStreamPartID } from '@streamr/protocol'
 import { Logger, collect, wait } from '@streamr/utils'
-import { Message } from '../Message'
+import { Message, convertStreamMessageToMessage } from '../Message'
 import { StreamrClientError } from '../StreamrClientError'
 import { Resends } from './Resends'
 
@@ -21,21 +21,21 @@ export const waitForStorage = async (
     }
     const macher = opts.messageMatchFn ?? ((msgTarget: Message, msgGot: Message) => (msgTarget.signature === msgGot.signature))
     const start = Date.now()
-    let last: Message[] | undefined
+    let last: StreamMessage[] | undefined
     let found = false
     while (!found) {
         const duration = Date.now() - start
         if (duration > opts.timeout) {
             logger.debug('Timed out waiting for storage to contain message', {
                 expected: message.streamMessage.getMessageID(),
-                lastReceived: last?.map((l) => l.streamMessage.getMessageID()),
+                lastReceived: last?.map((l) => l.getMessageID()),
             })
             throw new Error(`timed out after ${duration}ms waiting for message`)
         }
         const resendStream = await resends.resend(toStreamPartID(message.streamId, message.streamPartition), { last: opts.count })
         last = await collect(resendStream)
         for (const lastMsg of last) {
-            if (macher(message, lastMsg)) {
+            if (macher(message, convertStreamMessageToMessage(lastMsg))) {
                 found = true
                 logger.debug('Found matching message')
                 return
@@ -43,7 +43,7 @@ export const waitForStorage = async (
         }
         logger.debug('Retry after delay (matching message not found)', {
             expected: message.streamMessage.getMessageID(),
-            'last-3': last.slice(-3).map((l) => l.streamMessage.getMessageID()),
+            'last-3': last.slice(-3).map((l) => l.getMessageID()),
             delayInMs: opts.interval
         })
         await wait(opts.interval)