+++ packages/client/test/test-utils/fake/FakeNetworkNode.ts
@@ -1,8 +1,8 @@
-import { Lifecycle, scoped } from 'tsyringe'
-import pull from 'lodash/pull'
+import { NetworkNodeOptions, NodeId } from '@streamr/network-node'
 import { ProxyDirection, StreamMessage, StreamPartID } from '@streamr/protocol'
 import { MetricsContext } from '@streamr/utils'
-import { NodeId, NetworkNodeOptions } from '@streamr/network-node'
+import pull from 'lodash/pull'
+import { Lifecycle, scoped } from 'tsyringe'
 import { NetworkNodeFactory, NetworkNodeStub } from '../../../src/NetworkNodeFacade'
 import { FakeNetwork } from './FakeNetwork'
 
@@ -129,7 +129,7 @@ export class FakeNetworkNode implements NetworkNodeStub {
 @scoped(Lifecycle.ContainerScoped)
 export class FakeNetworkNodeFactory implements NetworkNodeFactory {
 
-    private network: FakeNetwork
+    private readonly network: FakeNetwork
 
     constructor(network: FakeNetwork) {
         this.network = network