+++ CHANGELOG.md
@@ -13,6 +13,8 @@ found [here](packages/broker/CHANGELOG.md).
 
 #### Added
 
+- Add optional config option `encryption.rsaKeyLength` to control the strength of RSA encryption in key-exchange (https://github.com/streamr-dev/network/pull/1505)
+
 #### Changed
 
 #### Deprecated